+++ source/standalone_examples/benchmarks/benchmark_robots_ur10.py
@@ -45,7 +45,7 @@ from functools import partial
 
 import carb
 import isaacsim.core.utils.stage as stage_utils
-import omni.physx as _physx
+import omni.physics.core
 import omni.timeline
 from isaacsim.core.api import PhysicsContext, World
 from isaacsim.core.prims import Articulation
@@ -119,7 +119,7 @@ def get_joint_commands(articulation_view, v_max, T, joint_indices):
     return position, velocity
 
 
-def on_physics_step(articulation_view, position_commands, velocity_commands, step):
+def on_physics_step(articulation_view, position_commands, velocity_commands, step, context):
     if position_commands is None:
         return
     timestep[0] += step
@@ -170,9 +170,8 @@ robot_view.initialize()
 omni.kit.app.get_app().update()
 
 position_commands, velocity_commands = get_joint_commands(robot_view, v_max, T, joint_indices)
-_physxIFace = _physx.get_physx_interface()
-physx_subscription = _physxIFace.subscribe_physics_step_events(
-    partial(on_physics_step, robot_view, position_commands, velocity_commands)
+physics_subscription = omni.physics.core.get_physics_simulation_interface().subscribe_physics_on_step_events(
+    pre_step=False, order=0, on_update=partial(on_physics_step, robot_view, position_commands, velocity_commands)
 )
 
 position_command = position_commands(0)