+++ source/extensions/isaacsim.gui.components/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.3.0] - 2025-10-17
+### Changed
+- Migrate PhysX subscription and simulation control interfaces to Omni Physics
+
 ## [1.2.1] - 2025-10-09
 ### Removed
 - Combo Box UI Wrapper (Duplicate from DropDown UI Wrapper)