+++ source/extensions/isaacsim.robot.policy.examples/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [4.2.0] - 2025-10-17
+### Changed
+- Migrate PhysX subscription and simulation control interfaces to Omni Physics
+
 ## [4.1.11] - 2025-07-07
 ### Fixed
 - Correctly enable omni.kit.loop-isaac in test dependency (fixes issue from 4.1.10)