+++ source/extensions/isaacsim.robot_setup.assembler/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [3.1.0] - 2025-10-17
+### Changed
+- Migrate PhysX subscription and simulation control interfaces to Omni Physics
+
 ## [3.0.11] - 2025-10-01
 ### Fixed
 - Issues with scenegraph instancing when fabric is disabled